% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradientGP.R, R/likelihoodGP.R,
%   R/metropolisGP.R
\name{gradientGP}
\alias{gradientGP}
\alias{gradientGPmatern}
\alias{posteriorgradientGPmatern}
\alias{gradientlogprior}
\alias{likelihoodGP}
\alias{likelihoodGPmatern}
\alias{posteriorGPmatern}
\alias{Gumbel}
\alias{PCrhomvar}
\title{Compute GP gradient}
\usage{
gradientGP(Xk, tau, h, nk, D)

gradientGPmatern(Xk, tau, h, nk, D, materncov, nu)

posteriorgradientGPmatern(Xk, tau, h, nk, D, materncov, nu, hyppar)

gradientlogprior(h, hyppar)

likelihoodGP(Xk, tau, h, nk, D)

likelihoodGPmatern(Xk, tau, h, nk, D, materncov, nu)

posteriorGPmatern(Xk, tau, h, nk, D, materncov, nu, hyppar)

Gumbel(x, lambda, log = TRUE)

PCrhomvar(rho, a, lambda1, lambda2, log = TRUE)

Gumbel(x, lambda, log = TRUE)

PCrhomvar(rho, a, lambda1, lambda2, log = TRUE)
}
\arguments{
\item{Xk}{The data}

\item{tau}{The indexing parameters}

\item{h}{GP hyperparameters}

\item{nk}{Number of observations}

\item{D}{number of samples}

\item{materncov}{\code{logical} indicating whether matern covariance is used}

\item{nu}{matern smoothness parameter.}

\item{hyppar}{prior hyperparameters of the penalised complexity prior.}

\item{x}{observation}

\item{lambda}{scale parameter of the type-2 Gumbel distribution}

\item{log}{\code{logical} indicating whether to return \code{log}. Default is \code{TRUE}}

\item{rho}{length-scale parameter}

\item{a}{amplitude}

\item{lambda1}{first parameter of distribution}

\item{lambda2}{second parameter of distribution}
}
\value{
Returns gp gradient

Returns gp gradient

Returns the gradient of the posterior

return the gradient of the log prior, length-scale, aamplitude and
noise

Returns gp negative log likelihood

Returns gp negative log likelihood

Returns the negative log posterior of the GP

Returns the likelihood of the type-2 GUmbel distribution

Returns the likelihood of the bivariate penalised complexity prior

Returns the likelihood of the type-2 GUmbel distribution

Returns the likelihood of the bivariate penalised complexity prior
}
\description{
Internal R function to pass R to C++, not for external use.

Internal R function to pass R to C++, not for external use.
}
