% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandle-utils.R
\name{meanOrganelle}
\alias{meanOrganelle}
\title{Computes Organelle means and variances using markers}
\usage{
meanOrganelle(object, fcol = "markers")
}
\arguments{
\item{object}{a instance of class \code{MSnset}}

\item{fcol}{a feature column indicating which feature define the markers}
}
\value{
returns a list of means and variances for each
}
\description{
Computes Organelle means and variances using markers
}
\examples{
library(pRolocdata)
data("tan2009r1")
meanOrganelle(object = tan2009r1)

}
