% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metropolisGP.R
\name{metropolisGP}
\alias{metropolisGP}
\alias{metropolisGPmatern}
\title{Perform metropolis update for GP hyperparameters}
\usage{
metropolisGP(
  inith,
  X,
  tau,
  nk,
  D,
  niter,
  hyperMean = c(0, 0, 0),
  hyperSd = c(1, 1, 1)
)

metropolisGPmatern(
  inith,
  X,
  tau,
  nk,
  D,
  niter,
  nu = 2,
  hyppar = c(1, 1, 1),
  propsd = c(0.3, 0.1, 0.1)
)
}
\arguments{
\item{inith}{initial hyperparamters}

\item{X}{The data}

\item{tau}{The indexing parameters}

\item{nk}{Number of observations}

\item{D}{number of samples}

\item{niter}{Number of MH iteractions}

\item{hyperMean}{A vector indicating the log-normal means. Default is \code{c(0,0,0)}.}

\item{hyperSd}{A vector indicating the log-normal standard deviations. Default is \code{c(1,1,1)}}

\item{nu}{Smoothness of the matern covariance}

\item{hyppar}{A vector indicating the penalised complexity prior hyperparameters.
Default is \code{c(1,1,1)}}

\item{propsd}{The proposal standard deviation. Default is \code{c(0.3,0.1,0.1)}. Do not
change unless you know what you are doing.}
}
\value{
Returns new hyperparamters and the acceptance rate
}
\description{
Function to perform Metropolis-Hastings for GP hyperparameters with different priors
}
