\name{bandle-package}
\alias{bandle-package}
\docType{package}
\title{
\packageTitle{bandle}
}
\description{
\packageDescription{bandle}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bandle}
\packageIndices{bandle}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{bandle}

Maintainer: \packageMaintainer{bandle}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }