% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandle-plots.R
\name{plotConvergence}
\alias{plotConvergence}
\title{Generates a histogram of ranks (a rank plot) for convergence}
\usage{
plotConvergence(params)
}
\arguments{
\item{params}{An instance of class \code{bandleParams}}
}
\value{
Returns the ranks of the number of outliers in each chain. The
side effect returns rank plots. Number of rank plots is equal to the number
of chains
}
\description{
Produces a rank plot to analyse convergence of MCMC algorithm
}
