% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandle-plots.R
\name{plotTable}
\alias{plotTable}
\title{Generates a table for visualising changes in
localisation between two conditions/datasets}
\usage{
plotTable(params, all = FALSE, fcol)
}
\arguments{
\item{params}{An instance of class \code{bandleParams} or an instance of
class \code{MSnSetList} of length 2.}

\item{all}{A logical specifying whether to count all proteins or only show
those that have changed in location between conditions. Default is
\code{FALSE}.}

\item{fcol}{If \code{params} is a \code{list} of \code{MSnSets}. Then
\code{fcol} must be defined. This is a \code{character} vector of length 2
to set different labels for each dataset. If only one label is specified,
and the \code{character} is of length 1 then this single label will be
used to identify the annotation column in both datasets.}
}
\value{
Returns a summary table of translocations of proteins between conditions.
}
\description{
Produces a table summarising differential localisation results
}
