% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandle-plots.R
\name{mcmc_plot_probs}
\alias{mcmc_plot_probs}
\alias{spatial2D}
\title{Generate a violin plot showing the probabilitiy of protein localisation
to different organelles}
\usage{
mcmc_plot_probs(
  params,
  fname,
  cond = 1,
  n = 1,
  bw = 0.05,
  scale = "width",
  trim = TRUE
)

spatial2D(
  object,
  params,
  fcol = "markers",
  dims = c(1, 2),
  cov.function = NULL,
  theta = 2,
  derivative = 2,
  k = 1,
  cond = 1,
  n = 1,
  breaks = c(0.99, 0.95, 0.9, 0.85, 0.8, 0.75, 0.7),
  aspect = 0.5
)
}
\arguments{
\item{params}{An instance of class \code{bandleParams}}

\item{fname}{The name of the protein to plot}

\item{cond}{Which conditions do we want to plot. Must be \code{1} or \code{2}. Default
is \code{1}}

\item{n}{The chain from which we plot the probability distribution. Default is 1.}

\item{bw}{The bandwidth use in probability distribution smoothing of geom_violin
Default is 0.05.}

\item{scale}{Scaling of geom_violin. Defaults to width.}

\item{trim}{trim parameter of geom_violin. Defaults to true.}

\item{object}{An instance of class \code{MSnSet} to provide the pca coordinates}

\item{fcol}{Feature columns that defines the markers. Defaults to "markers".}

\item{dims}{The PCA dimensions to plot. Defaults to \code{c(1,2)}}

\item{cov.function}{The covariance function for the smoothing kernel. Defaults
to wendland.cov}

\item{theta}{The theta parameter of the wendland.cov. Defaults to 2.}

\item{derivative}{The derivative paramter of the wendland.cov. Defaults to 2.}

\item{k}{The k parameter of the wendland.cov}

\item{breaks}{The levels at which to plot the contours. Defaults to c(0.99, 0.95, 0.9, 0.85, 0.8, 0.75, 0.7)}

\item{aspect}{The aspect ratio of the pca plots. Defaults to 0.5.}
}
\value{
returns a named vector of differential localisation probabilities

returns a named vector of differential localisation probabilities
}
\description{
These functions implement plotting functions for bandle objects
}
