% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{gpParams-class}
\alias{gpParams-class}
\alias{.gpParams}
\title{Container for GP results}
\description{
The \code{gpParams} infrastructure is used to store and process the GP results for
output from using the \code{fitGP} functions in \code{bandle}
}
\details{
Objects of the \code{gpParams} class are created with the \code{fitGP}, \code{fitGPmaternPC} or
\code{fitGPmatern} functions

These objects a list of posterior predictive means and standard deviations.
As well as maximum marginal likelihood for the GP
}
\section{Slots}{

\describe{
\item{\code{method}}{\code{character} indicating the GP method used}

\item{\code{M}}{A \code{list} of the posterior predictive means for each \code{K}
components of GPs fitted to the data}

\item{\code{sigma}}{A \code{numeric} of length \code{K} standard deviations fitted
to the data}

\item{\code{V}}{A \code{list} of the variance fitted to the data}

\item{\code{params}}{A \code{matrix} \code{array} of the MAP hyperparameters for the
GP}
}}

