% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gradientGPcpp}
\alias{gradientGPcpp}
\title{Computes gradients of  GP}
\usage{
gradientGPcpp(Xk, tau, h, nk, D)
}
\arguments{
\item{Xk}{The data}

\item{tau}{indexing term}

\item{h}{vector of hyperparamters}

\item{nk}{number of observations}

\item{D}{number of samples}
}
\description{
Computes gradients of  GP
}
