% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGP.R, R/fitGPmaternPC.R
\name{fitGP}
\alias{fitGP}
\alias{fitGPmaternPC}
\title{fit a Gaussian process to spatial proteomics data}
\usage{
fitGP(object = object, fcol = "markers")

fitGPmaternPC(
  object = object,
  fcol = "markers",
  materncov = TRUE,
  nu = 2,
  hyppar = c(1, 50, 50)
)
}
\arguments{
\item{object}{A instance of class \code{MSnSet}}

\item{fcol}{feature column to indicate markers. Default is "markers".}

\item{materncov}{\code{logical} indicating whether matern covariance is used}

\item{nu}{matern smoothness parameter. Default is 2.}

\item{hyppar}{The vector of penalised complexity hyperparameters}
}
\value{
means and standard deviation of gaussian process fitted to data,
along with hyperparameters with maximum marginal likelihood fitting. Side
effect, plot the markers with overlayed Gaussian processes.

returns a list of posterior predictive means and standard deviations.
As well as MAP hyperparamters for the GP. Side effect will plot the posterior
predictive overlayed with markers.
}
\description{
Helper function to fit GPs with squared exponential convariances,
maximum marginal likelihood

Function to fit matern GPs to data, side effect will plot posterior predictives
}
