% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{StatStratum}
\alias{StatStratum}
\title{inherits StatSratum}
\format{
An object of class \code{StatStratum} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
StatStratum
}
\description{
inherits StatSratum
}
\keyword{datasets}
