% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandle-plots.R
\name{mcmc_plot_probs}
\alias{mcmc_plot_probs}
\alias{spatial2D}
\alias{plotTranslocations}
\title{Generate a violin plot showing the probabilitiy of protein localisation
to different organelles}
\usage{
mcmc_plot_probs(
  params,
  fname,
  cond = 1,
  n = 1,
  bw = 0.05,
  scale = "width",
  trim = TRUE
)

spatial2D(
  object,
  params,
  fcol = "markers",
  dims = c(1, 2),
  cov.function = NULL,
  theta = 2,
  derivative = 2,
  k = 1,
  cond = 1,
  n = 1,
  breaks = c(0.99, 0.95, 0.9, 0.85, 0.8, 0.75, 0.7),
  aspect = 0.5
)

plotTranslocations(
  params,
  type = "alluvial",
  all = FALSE,
  fcol,
  col,
  labels = TRUE,
  labels.par = "adj",
  cex = 1,
  spacer = 4,
  table = FALSE,
  ...
)
}
\arguments{
\item{params}{An instance of class \code{bandleParams} or an instance of class
\code{MSnSetList} of length 2.}

\item{fname}{The name of the protein to plot}

\item{cond}{Which conditions do we want to plot. Must be \code{1} or \code{2}. Default
is \code{1}}

\item{n}{The chain from which we plot the probability distribution. Default is 1.}

\item{bw}{The bandwidth use in probability distribution smoothing of geom_violin
Default is 0.05.}

\item{scale}{Scaling of geom_violin. Defaults to width.}

\item{trim}{trim parameter of geom_violin. Defaults to true.}

\item{object}{An instance of class \code{MSnSet} to provide the pca coordinates}

\item{fcol}{If \code{params} is a \code{list} of \code{MSnSets}. Then \code{fcol} must
be defined. This is a \code{character} vector of length 2 to set different labels for
each dataset. If only one label is specified, and the \code{character} is of length 1
then this single label will be used to identify the annotation column in both datasets.}

\item{dims}{The PCA dimensions to plot. Defaults to \code{c(1,2)}}

\item{cov.function}{The covariance function for the smoothing kernel. Defaults
to wendland.cov}

\item{theta}{The theta parameter of the wendland.cov. Defaults to 2.}

\item{derivative}{The derivative paramter of the wendland.cov. Defaults to 2.}

\item{k}{The k parameter of the wendland.cov}

\item{breaks}{The levels at which to plot the contours. Defaults to c(0.99, 0.95, 0.9, 0.85, 0.8, 0.75, 0.7)}

\item{aspect}{The aspect ratio of the pca plots. Defaults to 0.5.}

\item{type}{A \code{character} specifying the type of visualisation to plot. One of
\code{"alluvial"} (default) or \code{"chord"}.}

\item{all}{A logical specifying whether to count all proteins or only show those that
have changed in location between conditions. Default is \code{FALSE}.}

\item{col}{A list of colours to define the classes in the data. If not defined then
the default \code{pRoloc} colours in \code{getStockCol()} are used.}

\item{labels}{Logical indicating whether to display class/organelle labels for the
chord segments or alluvial stratum. Default is \code{TRUE}.}

\item{labels.par}{If \code{type} is \code{"alluvial"}. Label style can be specified as
one of \code{"adj"}, \code{"repel"}. Default is \code{"adj"}.}

\item{cex}{Text size. Default is 1.}

\item{spacer}{A \code{numeric}. Default is 4. Controls the white space around the circos
plotting region.}

\item{table}{Logical. Print a summary table of translocations between subcellular classes.
Default is \code{FALSE}.}

\item{...}{Additional arguments passed to the \code{chordDiagram} function.}
}
\value{
returns a named vector of differential localisation probabilities

returns a named vector of differential localisation probabilities

Returns a directional circos/chord diagram showing the translocation of proteins
between conditions. If \code{type = "alluvial"} ouput is a \code{ggplot} object.
}
\description{
These functions implement plotting functions for bandle objects

Produces a chord diagram (circos plot) or an alluvial plot (also known as a Sankey diagram)
to show changes in location between two conditions or datasets.
}
